refactor(p2p_network/client): move conv for v0/1 to resemble sync handlers distribution
