+++ crates/pathfinder/src/p2p_network.rs
@@ -11,9 +11,6 @@ use tokio::sync::RwLock;
 use tracing::Instrument;
 
 pub mod client;
-#[allow(dead_code)]
-pub(crate) mod conv;
-#[allow(dead_code)]
 mod sync_handlers;
 
 // Silence clippy
